Remove unused tests code
Remove unused test mocks and utils. Reorganize tests code layout to
match main code layout(packages).
