+++ services/src/floodlight-modules/src/test/java/org/openkilda/floodlight/MockFloodlightProvider.java
@@ -29,7 +29,7 @@
 *    License for the specific language governing permissions and limitations
 *    under the License.
 **/
-package org.openkilda.floodlightcontroller.core.test;
+package org.openkilda.floodlight;
 
 import static org.junit.Assert.fail;
