+++ services/src/floodlight-modules/src/test/java/org/openkilda/floodlight/pathverification/PathVerificationPacketOutTest.java
@@ -31,7 +31,7 @@ import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.openkilda.floodlightcontroller.test.FloodlightTestCase;
+import org.openkilda.floodlight.FloodlightTestCase;
 import org.projectfloodlight.openflow.protocol.OFDescStatsReply;
 import org.projectfloodlight.openflow.protocol.OFPacketOut;
 import org.projectfloodlight.openflow.protocol.OFPortDesc;
