+++ services/src/floodlight-modules/src/test/java/org/openkilda/floodlight/test/standard/OutputCommands.java
@@ -13,15 +13,21 @@
  *   limitations under the License.
  */
 
-package org.openkilda.floodlight.message.command.encapsulation;
+package org.openkilda.floodlight.test.standard;
 
 import static java.util.Collections.singletonList;
 import static org.openkilda.floodlight.switchmanager.SwitchManager.DEFAULT_RULE_PRIORITY;
 import static org.openkilda.floodlight.switchmanager.SwitchManager.FLOW_COOKIE_MASK;
 import static org.openkilda.messaging.Utils.ETH_TYPE;
+import static org.projectfloodlight.openflow.protocol.OFMeterFlags.BURST;
+import static org.projectfloodlight.openflow.protocol.OFMeterFlags.KBPS;
+import static org.projectfloodlight.openflow.protocol.OFMeterModCommand.ADD;
 
 import net.floodlightcontroller.util.FlowModUtils;
+import org.openkilda.floodlight.OFFactoryMock;
+import org.projectfloodlight.openflow.protocol.OFFactory;
 import org.projectfloodlight.openflow.protocol.OFFlowAdd;
+import org.projectfloodlight.openflow.protocol.OFMeterMod;
 import org.projectfloodlight.openflow.protocol.match.MatchField;
 import org.projectfloodlight.openflow.types.EthType;
 import org.projectfloodlight.openflow.types.OFBufferId;
@@ -30,15 +36,68 @@ import org.projectfloodlight.openflow.types.OFVlanVidMatch;
 import org.projectfloodlight.openflow.types.U64;
 
 import java.util.Arrays;
+import java.util.HashSet;
 
-/**
- * Represent OF commands.
- * Code duplication is for more clear commands representation.
- */
-public class PushSchemeOutputCommands implements OutputCommands {
-    @Override
-    public OFFlowAdd ingressMatchVlanIdFlowMod(int inputPort, int outputPort, int inputVlan, int transitVlan,
-                                               long meterId, long cookie) {
+public interface OutputCommands {
+    OFFactory ofFactory = new OFFactoryMock();
+
+    OFFlowAdd egressReplaceFlowMod(int inputPort, int outputPort, int inputVlan, int outputVlan, long cookie);
+
+    OFFlowAdd egressPopFlowMod(int inputPort, int outputPort, int transitVlanId, long cookie);
+
+    OFFlowAdd egressPushFlowMod(int inputPort, int outputPort, int transitVlanId, int outputVlan, long cookie);
+
+    OFFlowAdd egressNoneFlowMod(int inputPort, int outputPort, int transitVlanId, long cookie);
+
+    OFFlowAdd ingressMatchVlanIdFlowMod(int inputPort, int outputPort, int inputVlan, int transitVlan,
+                                        long meterId, long cookie);
+
+    OFFlowAdd ingressNoMatchVlanIdFlowMod(int inputPort, int outputPort, int transitVlan,
+                                          long meterId, long cookie);
+
+    default OFFlowAdd ingressReplaceFlowMod(int inputPort, int outputPort, int inputVlan, int transitVlan,
+                                            long meterId, long cookie) {
+        return ingressMatchVlanIdFlowMod(inputPort, outputPort, inputVlan, transitVlan, meterId, cookie);
+    }
+
+    default OFFlowAdd ingressNoneFlowMod(int inputPort, int outputPort, int transitVlan, long meterId, long cookie) {
+        return ingressNoMatchVlanIdFlowMod(inputPort, outputPort, transitVlan, meterId, cookie);
+    }
+
+    default OFFlowAdd ingressPushFlowMod(int inputPort, int outputPort, int transitVlan, long meterId, long cookie) {
+        return ingressNoMatchVlanIdFlowMod(inputPort, outputPort, transitVlan, meterId, cookie);
+    }
+
+    default OFFlowAdd ingressPopFlowMod(int inputPort, int outputPort, int inputVlan, int transitVlan,
+                                        long meterId, long cookie) {
+        return ingressMatchVlanIdFlowMod(inputPort, outputPort, inputVlan, transitVlan, meterId, cookie);
+    }
+
+    default OFFlowAdd transitFlowMod(int inputPort, int outputPort, int transitVlan, long cookie) {
+        return ofFactory.buildFlowAdd()
+                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
+                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
+                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
+                .setBufferId(OFBufferId.NO_BUFFER)
+                .setPriority(DEFAULT_RULE_PRIORITY)
+                .setMatch(ofFactory.buildMatch()
+                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
+                        .setExact(MatchField.VLAN_VID, OFVlanVidMatch.ofVlan(transitVlan))
+                        .build())
+                .setInstructions(singletonList(
+                        ofFactory.instructions().applyActions(singletonList(
+                                ofFactory.actions().buildOutput()
+                                        .setMaxLen(0xFFFFFFFF)
+                                        .setPort(OFPort.of(outputPort))
+                                        .build()))
+                                .createBuilder()
+                                .build()))
+                .setXid(0L)
+                .build();
+    }
+
+    default OFFlowAdd oneSwitchReplaceFlowMod(int inputPort, int outputPort, int inputVlan, int outputVlan,
+                                              long meterId, long cookie) {
         return ofFactory.buildFlowAdd()
                 .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
                 .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
@@ -52,12 +111,9 @@ public class PushSchemeOutputCommands implements OutputCommands {
                 .setInstructions(Arrays.asList(
                         ofFactory.instructions().buildMeter().setMeterId(meterId).build(),
                         ofFactory.instructions().applyActions(Arrays.asList(
-                                ofFactory.actions().buildPushVlan()
-                                        .setEthertype(EthType.of(ETH_TYPE))
-                                        .build(),
                                 ofFactory.actions().buildSetField()
                                         .setField(ofFactory.oxms().buildVlanVid()
-                                                .setValue(OFVlanVidMatch.ofVlan(transitVlan))
+                                                .setValue(OFVlanVidMatch.ofVlan(outputVlan))
                                                 .build())
                                         .build(),
                                 ofFactory.actions().buildOutput()
@@ -68,12 +124,57 @@ public class PushSchemeOutputCommands implements OutputCommands {
                                 .build()))
                 .setXid(0L)
                 .build();
-
     }
 
-    @Override
-    public OFFlowAdd ingressNoMatchVlanIdFlowMod(int inputPort, int outputPort, int transitVlan,
-                                                 long meterId, long cookie) {
+    default OFFlowAdd oneSwitchNoneFlowMod(int inputPort, int outputPort, long meterId, long cookie) {
+        return ofFactory.buildFlowAdd()
+                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
+                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
+                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
+                .setBufferId(OFBufferId.NO_BUFFER)
+                .setPriority(DEFAULT_RULE_PRIORITY)
+                .setMatch(ofFactory.buildMatch()
+                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
+                        .build())
+                .setInstructions(Arrays.asList(
+                        ofFactory.instructions().buildMeter().setMeterId(meterId).build(),
+                        ofFactory.instructions().applyActions(singletonList(
+                                ofFactory.actions().buildOutput()
+                                        .setMaxLen(0xFFFFFFFF)
+                                        .setPort(OFPort.of(outputPort))
+                                        .build()))
+                                .createBuilder()
+                                .build()))
+                .setXid(0L)
+                .build();
+    }
+
+    default OFFlowAdd oneSwitchPopFlowMod(int inputPort, int outputPort, int inputVlan, long meterId, long cookie) {
+        return ofFactory.buildFlowAdd()
+                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
+                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
+                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
+                .setBufferId(OFBufferId.NO_BUFFER)
+                .setPriority(DEFAULT_RULE_PRIORITY)
+                .setMatch(ofFactory.buildMatch()
+                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
+                        .setExact(MatchField.VLAN_VID, OFVlanVidMatch.ofVlan(inputVlan))
+                        .build())
+                .setInstructions(Arrays.asList(
+                        ofFactory.instructions().buildMeter().setMeterId(meterId).build(),
+                        ofFactory.instructions().applyActions(Arrays.asList(
+                                ofFactory.actions().popVlan(),
+                                ofFactory.actions().buildOutput()
+                                        .setMaxLen(0xFFFFFFFF)
+                                        .setPort(OFPort.of(outputPort))
+                                        .build()))
+                                .createBuilder()
+                                .build()))
+                .setXid(0L)
+                .build();
+    }
+
+    default OFFlowAdd oneSwitchPushFlowMod(int inputPort, int outputPort, int outputVlan, long meterId, long cookie) {
         return ofFactory.buildFlowAdd()
                 .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
                 .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
@@ -86,40 +187,6 @@ public class PushSchemeOutputCommands implements OutputCommands {
                 .setInstructions(Arrays.asList(
                         ofFactory.instructions().buildMeter().setMeterId(meterId).build(),
                         ofFactory.instructions().applyActions(Arrays.asList(
-                                ofFactory.actions().buildPushVlan()
-                                        .setEthertype(EthType.of(ETH_TYPE))
-                                        .build(),
-                                ofFactory.actions().buildSetField()
-                                        .setField(ofFactory.oxms().buildVlanVid()
-                                                .setValue(OFVlanVidMatch.ofVlan(transitVlan))
-                                                .build())
-                                        .build(),
-                                ofFactory.actions().buildOutput()
-                                        .setMaxLen(0xFFFFFFFF)
-                                        .setPort(OFPort.of(outputPort))
-                                        .build()))
-                                .createBuilder()
-                                .build()))
-                .setXid(0L)
-                .build();
-
-    }
-
-    @Override
-    public OFFlowAdd egressPushFlowMod(int inputPort, int outputPort, int transitVlan, int outputVlan, long cookie) {
-        return ofFactory.buildFlowAdd()
-                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
-                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setBufferId(OFBufferId.NO_BUFFER)
-                .setPriority(DEFAULT_RULE_PRIORITY)
-                .setMatch(ofFactory.buildMatch()
-                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
-                        .setExact(MatchField.VLAN_VID, OFVlanVidMatch.ofVlan(transitVlan))
-                        .build())
-                .setInstructions(singletonList(
-                        ofFactory.instructions().applyActions(Arrays.asList(
-                                ofFactory.actions().popVlan(),
                                 ofFactory.actions().buildPushVlan()
                                         .setEthertype(EthType.of(ETH_TYPE))
                                         .build(),
@@ -138,83 +205,15 @@ public class PushSchemeOutputCommands implements OutputCommands {
                 .build();
     }
 
-    @Override
-    public OFFlowAdd egressPopFlowMod(int inputPort, int outputPort, int transitVlan, long cookie) {
-        return ofFactory.buildFlowAdd()
-                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
-                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setBufferId(OFBufferId.NO_BUFFER)
-                .setPriority(DEFAULT_RULE_PRIORITY)
-                .setMatch(ofFactory.buildMatch()
-                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
-                        .setExact(MatchField.VLAN_VID, OFVlanVidMatch.ofVlan(transitVlan))
-                        .build())
-                .setInstructions(singletonList(
-                        ofFactory.instructions().applyActions(Arrays.asList(
-                                ofFactory.actions().popVlan(),
-                                ofFactory.actions().popVlan(),
-                                ofFactory.actions().buildOutput()
-                                        .setMaxLen(0xFFFFFFFF)
-                                        .setPort(OFPort.of(outputPort))
-                                        .build()))
-                                .createBuilder()
-                                .build()))
-                .setXid(0L)
-                .build();
-    }
-
-    @Override
-    public OFFlowAdd egressNoneFlowMod(int inputPort, int outputPort, int transitVlan, long cookie) {
-        return ofFactory.buildFlowAdd()
-                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
-                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setBufferId(OFBufferId.NO_BUFFER)
-                .setPriority(DEFAULT_RULE_PRIORITY)
-                .setMatch(ofFactory.buildMatch()
-                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
-                        .setExact(MatchField.VLAN_VID, OFVlanVidMatch.ofVlan(transitVlan))
-                        .build())
-                .setInstructions(singletonList(
-                        ofFactory.instructions().applyActions(Arrays.asList(
-                                ofFactory.actions().popVlan(),
-                                ofFactory.actions().buildOutput()
-                                        .setMaxLen(0xFFFFFFFF)
-                                        .setPort(OFPort.of(outputPort))
-                                        .build()))
-                                .createBuilder()
-                                .build()))
-                .setXid(0L)
-                .build();
-    }
-
-    @Override
-    public OFFlowAdd egressReplaceFlowMod(int inputPort, int outputPort, int inputVlan, int outputVlan, long cookie) {
-        return ofFactory.buildFlowAdd()
-                .setCookie(U64.of(cookie & FLOW_COOKIE_MASK))
-                .setHardTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setIdleTimeout(FlowModUtils.INFINITE_TIMEOUT)
-                .setBufferId(OFBufferId.NO_BUFFER)
-                .setPriority(DEFAULT_RULE_PRIORITY)
-                .setMatch(ofFactory.buildMatch()
-                        .setExact(MatchField.IN_PORT, OFPort.of(inputPort))
-                        .setExact(MatchField.VLAN_VID, OFVlanVidMatch.ofVlan(inputVlan))
-                        .build())
-                .setInstructions(singletonList(
-                        ofFactory.instructions().applyActions(Arrays.asList(
-                                ofFactory.actions().popVlan(),
-                                ofFactory.actions().buildSetField()
-                                        .setField(ofFactory.oxms().buildVlanVid()
-                                                .setValue(OFVlanVidMatch.ofVlan(outputVlan))
-                                                .build())
-                                        .build(),
-                                ofFactory.actions().buildOutput()
-                                        .setMaxLen(0xFFFFFFFF)
-                                        .setPort(OFPort.of(outputPort))
-                                        .build()))
-                                .createBuilder()
-                                .build()))
+    default OFMeterMod installMeter(long bandwidth, long burstSize, long meterId) {
+        return ofFactory.buildMeterMod()
+                .setMeterId(meterId)
+                .setCommand(ADD)
+                .setMeters(singletonList(ofFactory.meterBands()
+                        .buildDrop()
+                        .setRate(bandwidth)
+                        .setBurstSize(burstSize).build()))
+                .setFlags(new HashSet<>(Arrays.asList(KBPS, BURST)))
                 .setXid(0L)
                 .build();
     }
